requirements.txt
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
pycodestyle==2.7.0
python-dotenv==0.19.0
requests==2.26.0
toml==0.10.2
urllib3==1.26.6
